backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.116.1",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-geohash>=0.8.5",
    "sqlalchemy>=2.0.43",
]
